refactor(carbon): extract dashboard tabs into widget files (Refs #563)
Moves the private `_ChartsTab` (ConsumerWidget) and `_AchievementsTab`
(StatelessWidget) out of `carbon_dashboard_screen.dart` and into their
own public widget files under `lib/features/carbon/presentation/widgets/`.
The screen drops from 348 LOC to 118 LOC, well under the 300-LOC target.

Pure refactor — no behavior change. Adds smoke tests for both new
public classes; existing carbon dashboard tests pass unchanged.
